+++ src/blocks/homepage-articles/edit.js
@@ -195,7 +195,10 @@ class Edit extends Component {
 						</h3>
 					) }
 					{ IS_SUBTITLE_SUPPORTED_IN_THEME && showSubtitle && (
-						<RawHTML key="subtitle" className="newspack-post-subtitle">
+						<RawHTML
+							key="subtitle"
+							className="newspack-post-subtitle newspack-post-subtitle--in-homepage-block"
+						>
 							{ post.meta.newspack_post_subtitle || '' }
 						</RawHTML>
 					) }