+++ src/blocks/homepage-articles/view.php
@@ -134,7 +134,7 @@ function newspack_blocks_get_homepage_articles_css_string( $attrs ) {
 
 	ob_start();
 	?>
-		.wpnbha .entry-title {
+		.wpnbha article .entry-title {
 			font-size: 1.2em;
 		}
 		.wpnbha .entry-meta {
@@ -193,14 +193,14 @@ function newspack_blocks_get_homepage_articles_css_string( $attrs ) {
 				}
 				if ( in_array( $scale, [ 1, 2, 3 ], true ) ) {
 					echo esc_html(
-						".wpnbha.ts-$scale article .newspack-post-subtitle,.entry-wrapper p,.entry-wrapper .more-link,.entry-meta {font-size: 0.8em;}"
+						".wpnbha.ts-$scale article .newspack-post-subtitle, .wpnbha.ts-$scale article .entry-wrapper p, .wpnbha.ts-$scale article .entry-wrapper .more-link, .wpnbha.ts-$scale article .entry-meta {font-size: 0.8em;}"
 					);
 				}
 			}
 		}
 		if ( isset( $attrs['showSubtitle'] ) && in_array( 1, $attrs['showSubtitle'], false ) ) { // phpcs:ignore WordPress.PHP.StrictInArray.FoundNonStrictFalse
 			echo esc_html(
-				'.newspack-post-subtitle {
+				'.newspack-post-subtitle--in-homepage-block {
 					margin-top: 0.3em;
 					margin-bottom: 0;
 					line-height: 1.4;
@@ -229,7 +229,7 @@ function newspack_blocks_render_block_homepage_articles( $attributes ) {
 	$block_name = apply_filters( 'newspack_blocks_block_name', 'newspack-blocks/homepage-articles' );
 
 	// Gather all Homepage Articles blocks on the page and output only the needed CSS.
-	// This CSS will be printed right after .entry-content.
+	// This CSS will be printed along with the first found block markup.
 	global $newspack_blocks_hpb_all_blocks;
 	$inline_style_html = '';
 	if ( ! is_array( $newspack_blocks_hpb_all_blocks ) ) {
@@ -358,6 +358,7 @@ function newspack_blocks_render_block_homepage_articles( $attributes ) {
 			class="<?php echo esc_attr( $classes ); ?>"
 			style="<?php echo esc_attr( $styles ); ?>"
 			>
+			<?php echo $inline_style_html; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
 			<div data-posts data-current-post-id="<?php the_ID(); ?>">
 				<?php if ( '' !== $attributes['sectionHeader'] ) : ?>
 					<h2 class="article-section-title">
@@ -404,7 +405,7 @@ function newspack_blocks_render_block_homepage_articles( $attributes ) {
 	$content = ob_get_clean();
 	Newspack_Blocks::enqueue_view_assets( 'homepage-articles' );
 
-	return $inline_style_html . $content;
+	return $content;
 }
 
 /**